requirements.txt
fastapi==0.101.0
firebase_admin==6.1.0
pydantic==1.10.8
python-dotenv==1.0.0
email-validator==2.0.0